feat(webcomp/button): Amélioration de la création
